backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
python-dotenv==1.0.0
elevenlabs==2.18.0
requests==2.31.0
mutagen==1.47.0
google-genai

